refactor: split doctor command facade
